+++ spec/unit/resources/matchers_spec.rb
@@ -27,6 +27,7 @@ describe 'oneview_test::default' do
     expect(chef_run).to_not remove_oneview_enclosure('')
     expect(chef_run).to_not refresh_oneview_enclosure('')
     expect(chef_run).to_not reconfigure_oneview_enclosure('')
+    expect(chef_run).to_not patch_oneview_enclosure('')
 
     # oneview_enclosure_group
     expect(chef_run).to_not create_oneview_enclosure_group('')