+++ CHANGELOG.md
@@ -23,6 +23,7 @@
     - Also supports SAS LIGs for Synergy in this logical_interconnect_groups property
   - volume support for API300
   - volume_template support for API300
+  - enclosure support for API300. Also added `:patch` action
 
 ### 1.1.0
   - Add support for client ENV variables